requirements.txt
boto3==1.26.22
botocore==1.29.22
cloudwatch==1.0.5
jmespath==1.0.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
